requirements.txt
fitz==0.0.1.dev2
nest_asyncio==1.6.0
spacy==3.8.2
transformers==4.47.0
python-dotenv==1.0.1
sentence_transformers==3.3.1
docling
pdfplumber
openai
llama-index-core 
llama_index==0.12.8
llama-index-readers-file
llama-index-readers-docling 
llama-index-node-parser-docling  
llama-index-llms-bedrock
llama-index-llms-azure-openai
llama_index-embeddings-azure-openai
azure-storage-blob
azure-search-documents
fastapi
fastapi[standard]
uvicorn

# uvicorn run:app --host 0.0.0.0 --port 8000 --workers 4